requirements.txt
Pillow
qrcode
PyMySQL
cryptography